requirements.txt
streamlit
tensorflow==2.13.0
numpy
Pillow